secure member and chat routes by returning data only to owners

+++ src/main/kotlin/com/example/controller/feature_member_manage/MemberRoutes.kt
@@ -1,11 +1,10 @@
 package com.example.controller.feature_member_manage
 
+import com.example.controller.util.isRequestedDataOwner
 import com.example.domain.dao.service.MemberService
 import com.example.domain.model.Member
 import io.ktor.http.*
 import io.ktor.server.application.*
-import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -15,26 +14,41 @@ import io.ktor.server.util.*
 fun Route.member(memberService: MemberService) {
     post("/member/add-member") {
         val member = call.receive<Member>()
-        memberService.createOrUpdateMember(member).also {
-            call.respond(HttpStatusCode.Created, it)
+        if (isRequestedDataOwner(member.uid)) {
+            memberService.createOrUpdateMember(member).also {
+                call.respond(HttpStatusCode.Created, it)
+            }
         }
+        call.respond(HttpStatusCode.Unauthorized)
     }
     put("/member/update-member-name") {
         val member = call.receive<Member>()
-        memberService.updateMemberName(member.uid, member.name).also {
-            call.respond(HttpStatusCode.OK)
+        if (isRequestedDataOwner(member.uid)) {
+            memberService.updateMemberName(member.uid, member.name).also {
+                call.respond(HttpStatusCode.OK)
+                return@put
+            }
         }
+        call.respond(HttpStatusCode.Unauthorized)
     }
     get("/member?id={id}") {
         val memberId = call.parameters.getOrFail("id")
-        memberService.getMember(memberId).also {
-            call.respond(HttpStatusCode.OK, it)
+        if (isRequestedDataOwner(memberId)) {
+            memberService.getMember(memberId).also {
+                call.respond(HttpStatusCode.OK, it)
+                return@get
+            }
         }
+        call.respond(HttpStatusCode.Unauthorized)
     }
     delete("/member?id={id}") {
         val memberId = call.parameters.getOrFail("id")
-        memberService.deleteMember(memberId).also {
-            call.respond(HttpStatusCode.NoContent)
+        if (isRequestedDataOwner(memberId)) {
+            memberService.deleteMember(memberId).also {
+                call.respond(HttpStatusCode.NoContent)
+                return@delete
+            }
         }
+        call.respond(HttpStatusCode.Unauthorized)
     }
 }
